refactor(broker): extract Broker orchestration into broker.rs
Moves the Broker struct, Broker::new and all its methods (build_provers,
fetch_and_upload_*, download_image, handle_join_result, start_service,
start_chain_pipeline, shutdown_and_cancel_critical_tasks), the
resolve_deployment helper plus its private validate_deployment_config
sibling, and the seven channel-capacity constants out of the 1500-line
lib.rs into their own module.

Public items are re-exported via `pub use broker::{Broker,
resolve_deployment}` so call-sites (broker::Broker, etc.) keep working.

lib.rs imports trimmed to drop pieces only used by the extracted code:
the alloy provider stack, NamedChain, BoundlessMarketService,
OrderStreamClient, StorageDownloader, ConfigWatcher/TelemetryMode,
SetVerifierService, ProverObj, JoinSet, broadcast/mpsc/RwLock,
CancellationToken, Url, plus the standard PathBuf/Arc/Duration that
weren't needed at the top level any more.

`pub(crate) use storage::ConfigurableDownloader;` is added in lib.rs
because proving.rs and order_pricer.rs reach the type via
crate::ConfigurableDownloader (the path that used to come from the
removed `use storage::ConfigurableDownloader;` line).

After this commit lib.rs is ~315 lines (down from ~1500 at the start of
PR-2): module declarations, top-level re-exports, BatchStatus /
AggregationState / Batch (still here), now_timestamp / format_expiries
/ is_dev_mode utilities, and the feature-gated test_utils::BrokerBuilder
that will move next.
